// src/app/iop/model/entity.response.ts
import {IopsignalRresponse} from './iopsignalRresponse';

/**
 * Sample model response for entity.
 */
export class EntityResponse {
  active_hrs_yesterday = 0;
  age: null;
  assignments: 0;
  breed: null;
  chassis_number: null;
  cnic: null;
  created_datetime: '2018-12-13 20:38:39';
  customer: 11;
  device_name: 631;
  device_name_method: 'homeappliance1';
  end_datetime: '';
  energy_consumed_yesterday: 1480;
  engine_number: null;
  entity_sub_type: 2001;
  entity_sub_type_name: 'Water Heater';
  errors_yesterday: 0;
  ethnicity: null;
  id: any;
  last_temperature: 0;
  last_updated: '';
  leased_owned: any;
  make: null;
  model: null;
  modified_by: any;
  modified_by_email: any;
  modified_by_name: any;
  modified_datetime: any;
  module: any;
  module_name: any;
  name: any;
  obd2_compliant: true;
  past_club: null;
  player_position: null;
  status: 1;
  status_label: any;
  tdist_last24Hrs: 0;
  total_distance: 0;
  tvol_last24Hrs: 0;
  type: 62;
  volume_capacity: null;
  weight: null;
  signalRresponse: IopsignalRresponse = new IopsignalRresponse(null, null, false, null, 0, null, null, null, null);

}